[camera] improve extension function visibility for consumption from java programming language
